requirements.txt
# Core Dependencies
face-recognition==1.3.0
dlib==19.24.2
opencv-python==4.6.0.66
customtkinter==5.2.0
numpy==1.26.4
pillow==10.1.0
pyyaml==6.0.1
pandas==2.1.4
python-dateutil==2.8.2
ffmpeg-python==0.2.0

# Support Libraries
face-recognition-models==0.3.0
setuptools==68.2.2
wheel==0.41.3
typing_extensions==4.9.0

# Desktop development with C++
# MSVC v143 - VS 2022 C++ x64/x86 build tools
# MSVC v142 - VS 2019 C++ x64/x86 build tools
# Windows 10 SDK (10.0.19041.0)
# C++ CMake tools for Windows
# C++ Clang Compiler for Windows (Recommended)
#  Windows Performance Toolkit
# C++ Build Insights


# Core Dependencies
# face-recognition (1.3.0) – Main library for facial recognition using deep learning.
# dlib (19.24.2) – A toolkit providing machine learning and image processing functionalities, essential for face recognition.
# opencv-python (4.6.0.66) – Open-source computer vision library for image and video processing.
# customtkinter (5.2.0) – A modern UI library for creating graphical user interfaces (GUIs) in Python.
# numpy (1.26.4) – A fundamental library for numerical computing and handling arrays.
# pillow (10.1.0) – Image processing library (an improved version of PIL).
# pyyaml (6.0.1) – Used for parsing and working with YAML configuration files.
# pandas (2.1.4) – Data manipulation and analysis library, useful for handling datasets.
# python-dateutil (2.8.2) – Extends Python’s datetime module for flexible date parsing.
# ffmpeg-python (0.2.0) – Python wrapper for FFmpeg, used for handling multimedia files (audio/video processing).

# Support Libraries
# face-recognition-models (0.3.0) – Pre-trained models required by face-recognition.
# setuptools (68.2.2) – Helps in packaging and distributing Python projects.
# wheel (0.41.3) – A package format that speeds up Python installations.
# typing_extensions (4.9.0) – Provides additional type hints for Python’s type system.